MNist_ttt4275/requirements.txt
python-mnist==0.7
